Move helpers.py in email_notifications to sender
Author thinks recipients list getter is more suitable to be in sender.

Refs TILA-2119
