+++ email_notification/tasks.py
@@ -2,8 +2,8 @@ from typing import List
 
 from django.conf import settings
 
-from email_notification.helpers import get_staff_notification_recipients
 from email_notification.models import EmailType
+from email_notification.sender.helpers import get_staff_notification_recipients
 from email_notification.sender.senders import send_reservation_email_notification
 from reservations.models import Reservation
 from tilavarauspalvelu.celery import app
